pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "physlean-bench"
version = "0.1.0"
description = "Benchmark construction and evaluation scaffold for Lean 4 physics theorem proving research."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "physlean-bench contributors" }
]
license = { text = "MIT" }
dependencies = [
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-cov>=5.0",
  "mypy>=1.8"
]
trace = [
  "lean-dojo>=4.20.0"
]

[project.scripts]
physlean-bench = "physlean_bench.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-q"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
